third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/video-intrinsic-width-height-expected.txt

This is a testharness.js-based test.
[PASS] no width/height attributes
[PASS] only width attribute
[PASS] only height attribute
[PASS] both width/height attributes
[FAIL] both width/height attributes and style
  assert_equals: height expected 300 but got 150
Harness: the test ran to completion.
